src/test/java/com/smartdubai/billingsystem/domain/CustomerTest.java
package com.smartdubai.billingsystem.domain;

import junit.framework.Assert;
import org.junit.Test;
import com.smartdubai.billingsystem.base.RetailBillingSystemBaseTest;

public class CustomerTest {
	
	@Test
	public void shouldReturnTrueForMoreThanTwoYearOldRegistration() {
		Customer customer = RetailBillingSystemBaseTest.getCustomerWithMoreThanTwoYearOldRegistration();
		Assert.assertTrue(customer.isAvailableForLoyaltyDiscount());
	}
	
	@Test
	public void shouldReturnFalseForNonEmployee() {
		Customer customer = RetailBillingSystemBaseTest.getAffiliateCustomer();
		Assert.assertFalse(customer.isAvailableForLoyaltyDiscount());
	}
	
	@Test
	public void shouldReturnFalseForLessThanTwoYearOldRegistration() {
		Customer customer = RetailBillingSystemBaseTest.getCustomerWithLessThanTwoYearOldRegistration();
		Assert.assertFalse(customer.isAvailableForLoyaltyDiscount());
	}
}
